requirements.txt
opencv-python==4.12.0.88
pillow==11.3.0
pyautogui==0.9.54
ruff==0.12.9
mss==10.0.0